% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R, R/html_dependency.R
\name{button}
\alias{button}
\alias{html_dependency_applause}
\title{Add an Applause Button to your page}
\usage{
button(
  ...,
  style = NULL,
  width = "50px",
  height = width,
  color = NULL,
  multiclap = FALSE,
  url = NULL,
  url_api = NULL
)

html_dependency_applause()
}
\arguments{
\item{...}{Additional attributes for the \verb{<applause-button>} tag}

\item{style}{Inline CSS styles}

\item{width, height}{The width and height of the applause button, for best
results the button should be square.}

\item{color}{The color of the button as a valid CSS color}

\item{multiclap}{Should users be allowed to clap more than once?}

\item{url}{The URL used to track total claps, if unset the current referring page URL will be used}

\item{url_api}{A URL for a custom Applause button back-end API, see
\url{https://github.com/ColinEberhardt/applause-button-server} for an example.}
}
\value{
An \code{\link[htmltools:tag]{htmltools::tagList()}} for an Applause button
}
\description{
Add an applause button to your page or Shiny app that allows users to
applaud your excellent work.
}
\section{Functions}{
\itemize{
\item \code{html_dependency_applause}: The applause button HTML dependency
}}

\references{
https://applause-button.com/
}
